Horrible mess
Works as a data scraper. Needs alot of cleaning up. Better way to find "go next page" button would be using cv2 to draw a mask of the ">" mark and check if it is found on the page.  Next commit also needs a crawler that checks job_tilte and employer from the scraped csv file and goes to the employers webpage and then sends the application.